requirements.txt
Django==1.8.8
Mezzanine==4.0.1
Pillow==3.1.0
Pygments==2.0.2
argparse==1.2.1
beautifulsoup4==4.4.1
bleach==1.4.2
chardet==2.3.0
django-contrib-comments==1.6.2
django-game-info==0.1
django-haystack==2.3.2
django-pagination-py3==1.1.1
djangorestframework==3.3.2
filebrowser-safe==0.4.0
future==0.15.2
grappelli-safe==0.4.1
html5lib==0.9999999
oauthlib==1.0.3
postmarkup==1.2.2
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
tzlocal==1.2
wsgiref==0.1.2
-e git://github.com/slav0nic/DjangoBB.git#egg=djangobb_forum
python-valve==0.1.1
python-social-auth==0.2.13
django-paypal==0.2.7
